+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: ataira <ataira@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/30 20:35:34 by ataira            #+#    #+#             */
-/*   Updated: 2023/01/31 01:11:51 by ataira           ###   ########.fr       */
+/*   Updated: 2023/02/03 00:42:02 by ataira           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	int		len;
-	int		i;
+	size_t	len;
+	size_t	i;
 	char	*str;
 
 	len = ft_strlen(s);
